refactor(B2B-707): put the catalogs in MySQL instead of Azure SQL
The catalogs QuoteCenter takes over from the legacy application — clubs, routes
and tariffs, fiscal document types, payment methods, exchange rates — now live
in the same MySQL database as everything else, not in a second Azure SQL
instance. The Azure server has been deleted.

What the split cost was never the T-SQL port. It was that two engines cannot be
joined, and two of our queries cross the boundary: DeliveryLogRepository joins
ps_delivery_log_cargue to quotation_delivery, PriceSmartPaymentService joins
ps_tienda to quotations. In one database those stay SQL; across two they become
application code that pages one side and matches it against the other. Being in
the same database as the legacy tables is not the same as being theirs — the
move that matters is off ps_*, and that still happens.

Fourteen tables of migration_quotecenter_schema.sql were already applied here,
countries among them, already holding the thirteen countries with `code` as the
ISO2. So this is the rest of a design that was written and half-landed. Three
columns are added to countries — price_includes_tax, default_language, tax_name
— and eight tables created.

The import is plain INSERT..SELECT, 130 lines, replacing a Java generator and
900 KB of statements that existed only because the two ends were different
engines. It also runs in seconds rather than hanging: 6076 exchange rates are
one statement here, not 6076 round trips.

Gone with the split: both datasource configurations, the mssql-jdbc dependency,
and the whole azure/ directory. Spring Boot autoconfigures the single datasource
again. The flag becomes quotecenter.catalogs.own-tables, which is what it always
described — which tables, not which server.

Everything that mattered survives. CatalogSource keeps both implementations and
the fallback. The identifiers that are persisted on quotations still survive the
move: the route llave as routes.code, the tender key verbatim, and the felid as
the explicit fiscal_document_types.id. So do the two fixes the comparison found,
route_prices.minimum_quantity and the payment ordering. One foreign key that was
impossible across databases — exchange_rates.created_by_user_id to users — is
real now.

docs/catalogs.md carries the reasoning that was in azure/README.md, including
the open question about price_includes_tax for the three countries whose legacy
value is empty, and Panama's unreachable subtotal swap.

Verified both ways against the real database: with the flag off and on, routes
and document types come back byte-identical, payment methods match on every
description and tender key in the same order, and the rest of the API is
unaffected. Counts match the Azure run exactly — 61 clubs, 434 routes, 1302
tariffs, 32 document types, 314 country payment methods, 6076 rates.

+++ api/src/main/java/com/dqs/api/catalog/model/ExchangeRate.java
@@ -38,13 +38,9 @@ public class ExchangeRate {
     @Column(name = "source", length = 100)
     private String source;
 
-    /**
-     * A `users.id` from the MySQL database. Deliberately a plain value with no
-     * association: the two live in different databases and no foreign key can
-     * span them.
-     */
+    /** A `users.id`, and a real foreign key — same database. */
     @Column(name = "created_by_user_id")
-    private Long createdByUserId;
+    private Integer createdByUserId;
 
     @Column(name = "created_at", insertable = false, updatable = false)
     private Instant createdAt;
